refactor: simplify parser logic and duplicated code
