--- src/event/index.ts
@@ -1,4 +1,0 @@
-export * from "./cloudevent";
-export * from "./validation";
-export * from "./v1";
-export * from "./v03";
