--- src/event/v1/index.ts
@@ -1,3 +1,0 @@
-export * from "./cloudevent";
-export * from "./spec";
-export * from "./schema";
