+++ src/index.ts
@@ -1,12 +1,7 @@
-import {
-  CloudEvent,
-  CloudEventV03,
-  CloudEventV03Attributes,
-  CloudEventV1,
-  CloudEventV1Attributes,
-  ValidationError,
-  Version,
-} from "./event";
+import { CloudEvent, Version } from "./event/cloudevent";
+import { ValidationError } from "./event/validation";
+// import {Version} from './event/'
+import { CloudEventV03, CloudEventV03Attributes, CloudEventV1, CloudEventV1Attributes } from "./event/interfaces";
 
 import { Emitter, Receiver, Mode, Protocol, TransportOptions } from "./transport";
 import { Headers, headersFor } from "./transport/http/headers";
