+++ src/transport/http/structured_receiver.ts
@@ -5,8 +5,8 @@ import { parserByContentType } from "../../parsers";
 import { structuredParsers as v1Parsers } from "./v1/parsers";
 import { structuredParsers as v03Parsers } from "./v03/parsers";
 import { isString, isBase64, ValidationError, isStringOrObjectOrThrow } from "../../event/validation";
-import { CloudEventV1, validateV1 } from "../../event/v1";
-import { CloudEventV03, validateV03 } from "../../event/v03";
+import { CloudEventV1, CloudEventV03 } from "../../event/interfaces";
+import { validateV1, validateV03 } from "../../event/spec";
 import CONSTANTS from "../../constants";
 
 /**
